requirements.txt
undetected_chromedriver
pandas
python-dotenv 